Enriched data: manual categories
I created a new version of the data file (avis_1834_rich), now also containing the manual tags - so we can use them as a benchmark for the tags we try to find via R.

Please note that the heading for the full text of the ad changed from "adcontent" to "text", because "adcontent" is the heading for the categories of adcontent. in the original file (and we will get more files of that kind in the future)

I duplicated tag_corpus.R to tag_corpus_rich.R to (1) incorporate these changes and (2) create a ids_by_manualcat.json according to the adcontent column.

Just not sure how to actually do the latter (see line 57ff)